adiciona sales person configuration e altera a classe sales person

+++ StocklyAPI/Models/SalesPerson.cs
@@ -6,7 +6,7 @@ namespace StocklyAPI.Models
     {
         public string FullName { get; set; }
         public int SalesCount { get; set; }
-        public decimal TotalSales { get; set; }
+        public int TotalSales { get; set; }
         public DateTime HireDate { get; set; }
         public bool IsActive { get; set; }
         public string ContactNumber { get; set; }
